# docs/swagger.yaml
basePath: /
definitions:
...
  domain.CreateProjectRequest:
    properties:
      color:
        type: string
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  domain.DailyStat:
    properties:
      activity_count:
        type: integer
      stat_date:
        description: YYYY-MM-DD
        type: string
      user_id:
        type: string
    type: object
...
  domain.Project:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        description: '''secondary'' | ''nextup'' | ''inprocess'' | ''completed'''
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.SyncUserParams:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
...
  domain.UpdateProjectRequest:
    properties:
      color:
        type: string
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  handler.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  handler.ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/handler.ErrorDetail'
    type: object
  handler.ProjectListResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      projects:
        items:
          $ref: '#/definitions/domain.Project'
        type: array
      total:
        type: integer
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/feaziest/kfdesktopbe
  description: High-performance REST API backend for Knowledge Faultorance (FeaziestFlow)
    desktop knowledge base.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Knowledge Faultorance (FeaziestFlow) API
  version: "1.0"
paths:
...
  /api/v1/projects:
    get:
      description: Retrieve paginated list of non-deleted projects for the authenticated
        user
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page limit (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ProjectListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List user workspace projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project owned by the authenticated user
      parameters:
      - description: Project creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create a new workspace project
      tags:
      - projects
  /api/v1/projects/{projectId}:
    delete:
      description: Marks a project as deleted without removing physical records
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Soft-delete project
      tags:
      - projects
    get:
      description: Retrieve single project details owned by the authenticated user
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Project'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update project title, description, color, or Kanban status
      parameters:
      - description: Project UUID
        in: path
        name: projectId
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update existing project
      tags:
      - projects
...
  /api/v1/users/contributions:
    get:
      description: Retrieve 365-day contribution activity stats for heatmaps
      parameters:
      - description: Year for stats (defaults to current year)
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.DailyStat'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get user daily activity contributions
      tags:
      - users
  /api/v1/users/me:
    get:
      description: Retrieve the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
  /api/v1/users/sync:
    post:
      consumes:
      - application/json
      description: Create or update user profile synchronized from Firebase Auth into
        PostgreSQL
      parameters:
      - description: User sync payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SyncUserParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Sync user profile
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your Firebase ID Token (or
      mock-token:<uid> in development).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"